Move (un)hashCharStats into the Characters composable
